separate ts and tsx code
